==> analysis/v1a/analysis/1_check_participants.R <==
library(here)
library(tidyverse)
library(jsonlite)
library(tidyjson)
library(rlang)
source("helper.R")

data_path <- here("..","data","processed", "kitty-act-v1a-alldata.csv")
prolific_path <- here("..","data","processed","prolific_export_67883d8fbfe27e2157999d3b.csv")
qualtrics_path <- here("..","data","processed","kitty-act-v1a_February 13, 2025_14.02.csv")

#### read in data ####
d <- read_csv(data_path)
prolific <- read_csv(prolific_path)
top_participants <- read_csv(here("..","data","processed", "top_participants.csv"))
qualtrics <- read_csv(qualtrics_path) %>%
  slice(-1:-2) %>% # remove the extra rows generated by qualtrics
  #remove test datasets
  filter(!(PROLIFIC_PID %in% c("{{%PROLIFIC_PID%}}"))) %>%
  filter(!is.na(PROLIFIC_PID))
#### handle/ compare to prolific submissions ####
unique_pavlovia_participants <- d %>%
  distinct(prolific_id,session_id,study_id,code, subject)

top_participants_vec <- as.character(top_participants$subject)
#top prolific ids
top_prolific <- d %>%
  distinct(prolific_id,session_id,study_id,code, subject) %>%
  filter(subject %in% top_participants$subject) %>%
  select(prolific_id) %>%
  mutate(bonus = 0.2)
write_csv(top_prolific,here("..","data","processed","top_prolific_ids.csv"))

prolific_submitted <- prolific %>%
  filter(Status=="APPROVED")

#### check for duplicated participant ids and completion codes

prolific_submitted %>% 
  group_by(`Completion code`) %>% 
  filter(n()>1)

prolific_submitted %>% 
  group_by(`Participant id`) %>% 
  filter(n()>1)
length(unique(prolific_submitted$`Participant id`))

#### comparing participants
setdiff(unique_pavlovia_participants$prolific_id,prolific_submitted$`Participant id`)
setdiff(prolific_submitted$`Participant id`,unique_pavlovia_participants$prolific_id)
## one prolific participant missing for unknown reasons

setdiff(unique_pavlovia_participants$prolific_id,qualtrics$PROLIFIC_PID)
setdiff(qualtrics$PROLIFIC_PID,unique_pavlovia_participants$prolific_id)

#### join
prolific_submitted <- prolific_submitted %>%
  left_join(unique_pavlovia_participants, by=c("Participant id"="prolific_id"))

prolific_submitted <- prolific_submitted %>%
  mutate(code_match = case_when(
    `Completion code` == code ~ 1,
    TRUE ~ 0
  ))
sum(prolific_submitted$code_match,na.rm=TRUE)
#one participant entered their prolific_id instead of a completion code
#followed up with a message that confirmed their response

#### counterbalancing check
#check number of sessions per id
sessions_per_id <- d %>%
  distinct(prolific_id,file_name) %>%
  group_by(prolific_id) %>%
  count()

condition_assignment <- d %>%
  distinct(prolific_id,stim_set,number_training_image_order_index) %>%
  group_by(stim_set,number_training_image_order_index) %>%
  summarize(
    n=n()
  )


condition_assignment_1 <- d %>%
  distinct(prolific_id,stim_set,number_training_image_order_index,catk_parameter_1) %>%
  group_by(stim_set,number_training_image_order_index,catk_parameter_1) %>%
  summarize(
    n=n()
  )

condition_assignment_2 <- d %>%
  distinct(prolific_id,stim_set,number_training_image_order_index,catk_parameter_2) %>%
  group_by(stim_set,number_training_image_order_index,catk_parameter_2) %>%
  summarize(
    n=n()
  )
