==> analysis/v1a/analysis/3_analysis.Rmd <==
---
title: "Kitty Act V1A Analysis"
author: "Martin Zettersten"
date: "XXX"
output:
  pdf_document: default
  html_document: default
---
  
```{r setup, warning=FALSE, message=F}
library(here)
library(tidyverse)
library(ggplot2)
library(viridisLite)
library(lme4)
library(ggstance)
library(cowplot)
library(AICcmodavg)
library(RColorBrewer)
library(jsonlite)
library(tidyjson)
library(rlang)
library(car)
library(nnet)
library(mlogit)
library(lmerTest)
library(assertthat)
library(knitr)
source("helper.R")

data_path <- here("..","data","processed", "kitty-act-v1a-alldata-anonymized.csv")
sampling_data_path <- here("..","data","processed", "kitty-act_v1a-sampling-data.csv")
test_data_path <- here("..","data","processed", "kitty-act_v1a-test-data.csv")
test_data_long_path <- here("..","data","processed", "kitty-act_v1a-test-data-long.csv")
write_path <- here("..","data","processed")
figure_path <- here("figures")
```

# Data Processing

## Read in data

```{r, warning=FALSE, message=F}
d <- read_csv(data_path) %>%
  mutate(number_training_images_order = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))
sampling_data <- read_csv(sampling_data_path) %>%
  mutate(number_training_image_order = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))
test_array_clean <- read_csv(test_data_path) %>%
  mutate(number_training_image_order_vector = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))
test_array_options_clean <- read_csv(test_data_long_path) %>%
  mutate(number_training_images_order = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))
```


## run a couple of tests on the data

```{r}
#check that we have all of the participants we expect
assert_that(length(unique(sampling_data$subject))==length(unique(test_array_clean$subject)))
assert_that(length(unique(filter(sampling_data, exclude_participant==0)$subject))==60)
```

## Summarize Test Data

```{r}
## shorter test representation
subj_test_accuracy <- test_array_clean %>%
  group_by(subject,number_training_images,trial_number,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options) %>%
  summarize(
    total_choices = n(),
    ground_truth_correct_choices = sum(test_choice_type_label_consistent),
    ground_truth_incorrect_choices = sum(test_choice_type_label_consistent==0),
    training_consistent_choices = sum(test_choice_type_training_consistent),
    training_inconsistent_choices = sum(test_choice_type_training_consistent==0),
    training_category_match_subordinate_choices = sum(test_choice_match_category==1 & test_choice_type == "subordinate"),
    training_category_match_basic_choices = sum(test_choice_match_category==1 & test_choice_type == "basic"),
    training_category_match_superordinate_choices = sum(test_choice_match_category==1 & test_choice_type == "superordinate"),
    training_category_match_subordinate_percent = training_category_match_subordinate_choices,
    training_category_match_basic_percent = training_category_match_basic_choices,
    training_category_match_superordinate_percent = training_category_match_superordinate_choices/2
  ) %>%
  mutate(
    round = case_when(
      trial_number %in% c(1,2,3) ~ 1,
      trial_number %in% c(4,5,6) ~ 2)
  )

subj_test_accuracy_long <- subj_test_accuracy %>%
  select(subject,number_training_images,trial_number,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options,training_category_match_subordinate_percent,training_category_match_basic_percent,training_category_match_superordinate_percent) %>%
  group_by(subject,number_training_images,trial_number,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options) %>%
  pivot_longer(cols=training_category_match_subordinate_percent:training_category_match_superordinate_percent,names_to = "choice_type",values_to = "percent_chosen") %>%
  mutate(
    round = case_when(
      trial_number %in% c(1,2,3) ~ 1,
      trial_number %in% c(4,5,6) ~ 2)
  )

overall_accuracy_by_condition <- subj_test_accuracy_long %>%
  ungroup() %>%
  group_by(number_training_images,current_category_training_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )

# looking at if 1- or 3-Item first has an effect

overall_accuracy_by_condition <- subj_test_accuracy_long %>%
  ungroup() %>%
  group_by(round,number_training_images,current_category_training_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )

# accuracy based on training condition (N, I, B) and ground-truth (same level or 1 level up)

overall_accuracy_label_level_by_condition <- subj_test_accuracy_long %>%
  ungroup() %>%
  group_by(number_training_images,current_category_training_level,current_category_label_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )

# again, looking at if 1- or 3-Item first has an effect on above accuracy

overall_accuracy_label_level_by_condition_round <- subj_test_accuracy_long %>%
  ungroup() %>%
  group_by(round,number_training_images,current_category_training_level,current_category_label_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )


### longer test representation
total_number_options <- 12
subj_test_accuracy_all <- test_array_options_clean %>%
  group_by(subject,number_training_images,trial_number,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options_label,total_number_correct_options_training) %>%
  summarize(
    accuracy_training = mean(is_match_to_training, na.rm=TRUE),
    accuracy_label = mean(is_match_to_label,na.rm=TRUE),
    hit_rate_training = mean(hit_training,na.rm=TRUE),
    hit_rate_label = mean(hit_label,na.rm=TRUE),
    false_alarm_rate_training = mean(false_alarm_training,na.rm=TRUE),
    false_alarm_rate_label = mean(false_alarm_label,na.rm=TRUE)
  ) %>%
  mutate(
    hit_rate_training_adj= case_when(
      hit_rate_training==1 ~ 1 - 1/(2*total_number_options),
      hit_rate_training==0 ~ 1/(2*total_number_options),
      TRUE ~ hit_rate_training
    ),
    hit_rate_label_adj= case_when(
      hit_rate_label==1 ~ 1 - 1/(2*total_number_options),
      hit_rate_label==0 ~ 1/(2*total_number_options),
      TRUE ~ hit_rate_label
    ),
    false_alarm_rate_training_adj= case_when(
      false_alarm_rate_training==0 ~ 1/(2*total_number_options),
      false_alarm_rate_training==1 ~ 1 - 1/(2*total_number_options),
      TRUE ~ false_alarm_rate_training
    ),
    false_alarm_rate_label_adj= case_when(
      false_alarm_rate_label==0 ~ 1/(2*total_number_options),
      false_alarm_rate_label==1 ~ 1 - 1/(2*total_number_options),
      TRUE ~ false_alarm_rate_label
    )
  ) %>%
  mutate(
    dprime_training=qnorm(hit_rate_training_adj) - qnorm(false_alarm_rate_training_adj),
    dprime_label=qnorm(hit_rate_label_adj) - qnorm(false_alarm_rate_label_adj),
    c_training=-.5*(qnorm(hit_rate_training_adj) + qnorm(false_alarm_rate_training_adj)),
    c_label=-.5*(qnorm(hit_rate_label_adj) + qnorm(false_alarm_rate_label_adj)))

subj_test_accuracy_all <- subj_test_accuracy_all %>%
  left_join(sampling_data)

# gives accuracy based on training level, ground-truth level, and if sampling was confirm/explore

overall_accuracy_label_level_sampling_all <- subj_test_accuracy_all %>%
  ungroup() %>%
  group_by(sampling_choice_type,current_category_training_level,current_category_label_level) %>%
  summarize(
    N=n(),
    average_dprime_training=mean(dprime_training),
    average_dprime_label=mean(dprime_label)
  ) 
```

## Rank subject-level accuracy

```{r}
subj_test_rank <- subj_test_accuracy_all %>%
  ungroup() %>%
  filter(exclude_participant==0) %>%
  select(subject,number_training_images,current_category_training_level,current_category_label_level, accuracy_label,accuracy_training,dprime_label,dprime_training) %>%
  group_by(number_training_images,current_category_training_level,current_category_label_level) %>%
  arrange(-dprime_label,-dprime_training,.by_group = TRUE) %>%
  mutate(rank = rank(-dprime_label, ties.method="min"))

# median subject rank
subj_test_summary <- subj_test_rank %>%
  ungroup() %>%
  group_by(subject) %>%
  summarize(
    n=n(),
    avg_accuracy_label = mean(accuracy_label),
    avg_accuracy_training = mean(accuracy_training),
    avg_dprime_label = mean(dprime_label),
    avg_dprime_training = mean(dprime_training),
    median_rank = median(rank),
    mean_rank = mean(rank)
  )

ggplot(subj_test_summary,aes(avg_dprime_label,median_rank))+geom_point()+geom_smooth()
ggplot(subj_test_summary,aes(avg_dprime_label,mean_rank))+geom_point()+geom_smooth()

#use mean rank to get the top 15 participants
top_participants <- subj_test_summary %>%
  ungroup() %>%
  arrange(mean_rank) %>%
  slice_min(order_by=mean_rank,n=15) %>%
  select(subject)
write_csv(top_participants,here(write_path,"top_participants.csv"))
  
```


## Sampling

### Models

#### H1

Do learners flexibly shift their sampling choices depending on the training condition? 

```{r}
sampling_data <- sampling_data %>%
  mutate(
    sampled_category_level_kind_info_choice_order = case_when(
      sampled_category_level_kind_info == "outside_category" ~ "a_outside_category",
      sampled_category_level_kind_info == "sub" ~ "b_sub",
      sampled_category_level_kind_info == "bas" ~ "c_bas",
      sampled_category_level_kind_info == "sup" ~ "d_sup"
  ),
  current_category_training_level_order = case_when(
      current_category_training_level == "narrow" ~ "a_narrow",
      current_category_training_level == "intermediate" ~ "b_intermediate",
      current_category_training_level == "broad" ~ "c_broad"
  )
  )

sampling_data_3_images_only <- sampling_data %>%
  filter(number_training_images==3)

#fit multinomial model
nnetFixedModelStr <- 'sampled_category_level_kind_info_choice_order ~ current_category_training_level_order'
nnetFixedFit <-  multinom(as.formula(nnetFixedModelStr), sampling_data_3_images_only, maxit=10000, abstol=1e-8, reltol=1e-10)
summary(nnetFixedFit)
Anova(nnetFixedFit,type="III")

# Alternative (equivalent) approach with mlogit
# reformat data for mlogit
mlogit_sampling_data <-  mlogit.data(data=select(sampling_data_3_images_only,subject,sampled_category_level_kind_info_choice_order,current_category_training_level_order,current_category_kind), choice='sampled_category_level_kind_info_choice_order', shape="wide", id.var='subject')

# specify model formula
mlogit_sampling_formula_baseline <- ' sampled_category_level_kind_info_choice_order ~ 0 | 1 | 0'
mlogit_sampling_formula_training <- 'sampled_category_level_kind_info_choice_order ~ 0 | current_category_training_level_order | 0'

# fit models
mlogit_sampling_baseline <- mlogit(as.formula(mlogit_sampling_formula_baseline), data=mlogit_sampling_data, panel=FALSE, iterlim=10000, tol = 1e-8)
print(summary(mlogit_sampling_baseline))
mlogit_sampling_training <- mlogit(as.formula(mlogit_sampling_formula_training), data=mlogit_sampling_data, panel=FALSE, iterlim=10000, tol = 1e-8)
print(summary(mlogit_sampling_training))

# overall test of condition (identical result to nnet method above)
lrtest(mlogit_sampling_baseline,mlogit_sampling_training)
```

#### 1-Narrow vs. 3-Narrow

```{r}
#filter relevant conditions
sampling_data_1_vs_3 <- sampling_data %>%
  filter(current_category_training_level=="narrow") %>%
  mutate(
    number_training_images_c=(number_training_images-2)/2
  )

#fit multinomial model
nnetFixedModelStr <- 'sampled_category_level_kind_info_choice_order ~ number_training_images_c'
nnetFixedFit <-  multinom(as.formula(nnetFixedModelStr), sampling_data_1_vs_3, maxit=10000, abstol=1e-8, reltol=1e-10)
summary(nnetFixedFit)
Anova(nnetFixedFit,type="III")
```


### Robustness Checks

#### Controlling for Category Type

```{r}
nnetFixedModelStr <- 'sampled_category_level_kind_info_choice_order ~ current_category_training_level_order+current_category_kind'
nnetFixedFit <-  multinom(as.formula(nnetFixedModelStr), sampling_data_3_images_only, maxit=10000, abstol=1e-8, reltol=1e-10)
summary(nnetFixedFit)
Anova(nnetFixedFit,type="II")
```

#### Adding by-participant random effects

```{r}
#set up random effects and model fitting parameters
parNames=names(mlogit_sampling_training$coefficients)[1:3]
print(parNames)
rparArg=rep('n',length(parNames))
names(rparArg)=parNames
print(rparArg)
NRCRDRAWS=1000

#fit models
if (NRCRDRAWS>0){
mlogit_sampling_baseline_re=mlogit(as.formula(mlogit_sampling_baseline), data=mlogit_sampling_data, rpar=rparArg, 
                panel=TRUE, R=NRCRDRAWS,print.level=1)
print(summary(mlogit_sampling_baseline_re))
}
if (NRCRDRAWS>0){
mlogit_sampling_training_re=mlogit(as.formula(mlogit_sampling_formula_training), data=mlogit_sampling_data, rpar=rparArg, 
                panel=TRUE, R=NRCRDRAWS,print.level=1)
print(summary(mlogit_sampling_training_re))
}

# LR Test
lrtest(mlogit_sampling_baseline_re,mlogit_sampling_training_re)
```

### Plot

```{r}
sampling_data$sampled_category_level_kind_info_ord <- factor(sampling_data$sampled_category_level_kind_info,
                                                             levels=c("sub","bas","sup","outside_category"),
                                                             labels=c("subordinate","basic","superordinate","outside category"))
sampling_data$current_category_training_level_ord <- factor(sampling_data$current_category_training_level,levels=c("narrow","intermediate","broad"))

ggplot(sampling_data,aes(current_category_training_level_ord,fill=sampled_category_level_kind_info_ord))+
  geom_bar(position="fill")+
  scale_fill_brewer(name="Sampling Choice Type",palette="RdYlBu")+
  xlab("Training Condition")+
  ylab("Proportion of Sampling Choices")+
  facet_wrap(~number_training_images)#+
  # theme_cowplot(font_size=24)+
  # theme(axis.title = element_text(face="bold", size=20),
  #       axis.text.x  = element_text(angle=90, hjust=1,vjust=0.4))

## Sampling Choices - individual bars
sampling_data_long_structure <- expand.grid(
  subject=unique(sampling_data$subject),
  trial_number=seq(1,6),
  sampling_choice_type = c("subordinate","basic","superordinate","outside category")
)
sampling_data_long <-  sampling_data %>%
  group_by(subject,number_training_images,trial_number, current_category_training_level) %>%
  summarize(sampling_response = sampled_category_level_kind_info_ord) %>%
  arrange(subject,trial_number)
sampling_data_long_structure <- sampling_data_long_structure %>%
  left_join(sampling_data_long) %>%
  mutate(
    sampling_choice_chosen = case_when(
      sampling_response == sampling_choice_type ~ 1,
      TRUE ~ 0
    )
  )

sampling_data_long_structure_summarized <- sampling_data_long_structure %>%
  group_by(number_training_images,current_category_training_level, sampling_choice_type) %>%
  summarize(
    N=n(),
    sampling_response_avg=mean(sampling_choice_chosen,na.rm=TRUE),
    sampling_response_total=sum(sampling_choice_chosen,na.rm=TRUE),
    sampling_response_binom_lower=binom.test(sampling_response_total,N)$conf.int[1],
    sampling_response_binom_upper=binom.test(sampling_response_total,N)$conf.int[2],
  )
sampling_data_long_structure_summarized$current_category_training_level_ord <- factor(sampling_data_long_structure_summarized$current_category_training_level,levels=c("narrow","intermediate","broad"))


ggplot(sampling_data_long_structure_summarized,aes(current_category_training_level_ord,sampling_response_avg, group=sampling_choice_type,fill=sampling_choice_type))+
  geom_errorbar(aes(ymin=sampling_response_binom_lower, ymax=sampling_response_binom_upper),width=0.3, color="black",position=position_dodge(width=0.9))+
  geom_bar(stat="identity",position=position_dodge())+
  #theme_cowplot(font_size=30)+
  scale_fill_brewer(name="Sampling Choice Type",palette="RdYlBu")+
  scale_color_brewer(name="Sampling Choice Type",palette="RdYlBu")+
  xlab("Training Condition")+
  ylab("Proportion of Sampling Choices")+
  facet_wrap(~number_training_images)
```

## Test

### Plot

```{r}
## Test Plot
subj_test_prop <- test_array_options_clean %>%
  mutate(training_condition = case_when(
    number_training_images == 1 ~ "1-Item",
    TRUE ~ paste0("3-Item ", stringr::str_to_title(current_category_training_level))
  )) %>%
  ungroup() %>%
  group_by(subject,number_training_images,current_category_training_level,training_condition,current_category_label_level,current_category_kind) %>%
  summarize(
    N_subordinate = sum(test_image_type=="subordinate" & test_image_match_category==1),
    prop_subordinate = sum(test_image_type=="subordinate" & test_image_match_category==1 & is_chosen == 1)/sum(test_image_type=="subordinate" & test_image_match_category==1),
    N_basic = sum(test_image_type=="basic" & test_image_match_category==1),
    prop_basic = sum(test_image_type=="basic" & test_image_match_category==1 & is_chosen == 1)/sum(test_image_type=="basic" & test_image_match_category==1),
    N_superordinate = sum(test_image_type=="superordinate" & test_image_match_category==1),
    prop_superordinate = sum(test_image_type=="superordinate" & test_image_match_category==1 & is_chosen == 1)/sum(test_image_type=="superordinate" & test_image_match_category==1)
  )

subj_test_prop_long <- subj_test_prop %>%
  pivot_longer(N_subordinate:prop_superordinate,names_to=c(".value","choice_type"),names_sep="\\_")

overall_test_prop <- subj_test_prop %>%
  group_by(training_condition,number_training_images,current_category_training_level) %>%
  summarize(
    N=n(),
    subordinate_percent = mean(prop_subordinate),
    basic_percent = mean(prop_basic),
    superordinate_percent = mean(prop_superordinate)
  )

overall_test_prop_long <- subj_test_prop_long %>%
  group_by(training_condition,number_training_images,current_category_training_level,choice_type) %>%
  summarize(
    N=n(),
    avg_prop = mean(prop),
    prop_ci=qt(0.975, N-1)*sd(prop,na.rm=T)/sqrt(N),
    prop_lower_ci=avg_prop-prop_ci,
    prop_upper_ci=avg_prop+prop_ci,
  )
overall_test_prop_long$choice_type_ord <- factor(overall_test_prop_long$choice_type,levels=c("subordinate","basic","superordinate"))
overall_test_prop_long$training_condition_ord <- factor(overall_test_prop_long$training_condition,levels=c("1-Item","3-Item Narrow", "3-Item Intermediate","3-Item Broad"))

ggplot(overall_test_prop_long,aes(choice_type_ord,avg_prop,fill=choice_type_ord))+
  geom_bar(stat="identity")+
  geom_errorbar(aes(ymin=prop_lower_ci,ymax=prop_upper_ci),width=0.1)+
  facet_wrap(~training_condition_ord)+
  theme_cowplot(font_size=12)+
  scale_fill_manual(
    name="Test Choice Type",
    values=c("#d7191c","#fdae61","#abd9e9"))+
  xlab("Test Choice Type")+
  ylab("Average Proportion\nof Test Choices")+
  theme(legend.position="none")+
  theme(axis.title = element_text(face="bold", size=14),
        #axis.text.x  = element_text(angle=90, hjust=1,vjust=0.4)
        )
ggsave(here(figure_path,"kitty-act-test.png"),width=6,height=6)
  
```

## Sampling 


## Exploratory

### investigate the proportion of exploratory choices

```{r}
sampling_data <- sampling_data %>%
  mutate(is_exploratory=ifelse(sampling_choice_type=="explore",1,0)) %>%
  mutate(training_condition = case_when(
    number_training_images == 1 ~ "1-Item",
    TRUE ~ paste0("3-Item ", stringr::str_to_title(current_category_training_level))
  )) 

subj_exploratory <- sampling_data %>%
  group_by(subject) %>%
  summarize(
    N=n(),
    prop_exploratory=mean(is_exploratory)
  )

#exploratory averages by age group
overall_exploratory_age_group <- subj_exploratory %>%
  group_by() %>%
  summarize(
    N=n(),
    mean_explore=mean(prop_exploratory),
    sd(prop_exploratory,na.rm=T)/sqrt(N),
    se = sd(prop_exploratory,na.rm=T)/sqrt(N),
    ci=qt(0.975, N-1)*se,
    lower_se = mean_explore-se,
    upper_se = mean_explore+se,
    lower_ci=mean_explore-ci,
    upper_ci=mean_explore+ci,
  )


#model
m <- glmer(is_exploratory~ training_condition+(1|subject),data=sampling_data,family="binomial")
summary(m)
Anova(m,type="III")
```

## H4: Test Choices

## Training condition modulated participants’ choices at test (Proportion choices)

```{r}
# fit individual lmer models
## subordinate choices
m_sub <- lmer(prop_subordinate~training_condition+(1|subject),data=subj_test_prop)
summary(m_sub)
Anova(m_sub,type="III")

## basic choices
m_bas <- lmer(prop_basic~training_condition+(1|subject),data=subj_test_prop)
summary(m_bas)
Anova(m_bas,type="III")

## superordinate choices
m_sup <- lmer(prop_superordinate~training_condition+(1|subject),data=subj_test_prop)
summary(m_sup)
Anova(m_sup,type="III")

```

### H5: Proportion basic-level choices

#### Test

```{r}
subj_test_prop <- subj_test_prop %>%
  mutate(number_training_images_c=(number_training_images-2)/2)

m <- glmer(prop_basic ~ number_training_images_c+(1|subject)+(1|current_category_kind),data=filter(subj_test_prop,current_category_training_level=="narrow"),family="binomial")
summary(m)
#replicates the basic Xu/ Tenenbaum effect
```
